+++ src/main/java/net/sberg/elbook/glossarcmpts/GlossarController.java
@@ -17,46 +17,24 @@ package net.sberg.elbook.glossarcmpts;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import net.sberg.elbook.authcomponents.AuthUserDetails;
-import net.sberg.elbook.batchjobcmpts.BatchJob;
 import net.sberg.elbook.common.AbstractWebController;
-import net.sberg.elbook.common.ICommonConstants;
-import net.sberg.elbook.common.StringUtils;
-import net.sberg.elbook.jdbc.DaoPlaceholderProperty;
-import net.sberg.elbook.jdbc.DaoProjectionBean;
-import net.sberg.elbook.jdbc.JdbcGenericDao;
-import net.sberg.elbook.logeintragcmpts.LogEintrag;
-import net.sberg.elbook.logeintragcmpts.LogEintragArtikel;
-import net.sberg.elbook.mandantcmpts.*;
-import net.sberg.elbook.reportcmpts.Report;
-import net.sberg.elbook.stammdatenzertimportcmpts.VerzeichnisdienstImportCommandContainer;
-import net.sberg.elbook.stammdatenzertimportcmpts.VerzeichnisdienstImportErgebnis;
-import org.jboss.aerogear.security.otp.Totp;
-import org.jboss.aerogear.security.otp.api.Base32;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.util.FileSystemUtils;
 import org.springframework.web.bind.annotation.*;
 
-import java.io.File;
-import java.io.UnsupportedEncodingException;
-import java.net.URLEncoder;
-import java.time.LocalDateTime;
 import java.util.*;
 
 @Controller
 @RequiredArgsConstructor
 @Slf4j
 public class GlossarController extends AbstractWebController {
+
+    private static final String SEARCHTYPE_TID = "telematikId";
+    private static final String SEARCHTYPE_PID = "professionOID";
+    private static final String SEARCHTYPE_HOLDER = "holder";
 
     private final GlossarService glossarService;
 
@@ -68,31 +46,106 @@ public class GlossarController extends AbstractWebController {
 
     @RequestMapping(value = "/glossar/uebersicht", method = RequestMethod.POST)
     @ResponseStatus(value = HttpStatus.OK)
-    public String uebersicht(Model model, String telematikId) throws Exception {
-        if (telematikId.equals("")) {
-            model.addAttribute("telematikIdInfo", null);
-            model.addAttribute("verfuegbar", false);
-            model.addAttribute("suche", false);
+    public String uebersicht(Model model, String searchValue, String searchType) throws Exception {
+        if (searchType.equals(SEARCHTYPE_TID)) {
+            if (searchValue.equals("")) {
+                model.addAttribute("telematikIdInfo", null);
+                model.addAttribute("telematikIdInfoVerfuegbar", false);
+            }
+            else {
+                TelematikIdInfo telematikIdInfo = glossarService.getTelematikIdInfo(searchValue);
+                model.addAttribute("telematikIdInfo", telematikIdInfo);
+                model.addAttribute("telematikIdInfoVerfuegbar", telematikIdInfo != null);
+            }
+        }
+        else if (searchType.equals(SEARCHTYPE_PID)) {
+            model.addAttribute("telematikIdInfoVerfuegbar", false);
+            if (searchValue.equals("")) {
+                model.addAttribute("professionOIDInfos", new ArrayList<>());
+            }
+            else if (searchValue.equals(GlossarService.SEARCHVALUE_ALL_ELEMS)) {
+                List<ProfessionOIDInfo> professionOIDInfos = glossarService.getAllProfessionOIDInfo();
+                model.addAttribute("professionOIDInfos", professionOIDInfos);
+            }
+            else {
+                ProfessionOIDInfo professionOIDInfo = glossarService.getProfessionOIDInfo(searchValue);
+                model.addAttribute("professionOIDInfos", professionOIDInfo != null?List.of(professionOIDInfo):new ArrayList<>());
+            }
+
+        }
+        else if (searchType.equals(SEARCHTYPE_HOLDER)) {
+            model.addAttribute("telematikIdInfoVerfuegbar", false);
+            if (searchValue.equals("")) {
+                model.addAttribute("holderInfos", new ArrayList<>());
+            }
+            else if (searchValue.equals(GlossarService.SEARCHVALUE_ALL_ELEMS)) {
+                List<HolderInfo> holderInfos = glossarService.getAllHolderInfo();
+                model.addAttribute("holderInfos", holderInfos);
+            }
+            else {
+                HolderInfo holderInfo = glossarService.getHolderInfo(searchValue);
+                model.addAttribute("holderInfos", holderInfo != null?List.of(holderInfo):new ArrayList<>());
+            }
         }
         else {
-            TelematikIdInfo telematikIdInfo = glossarService.get(telematikId);
-            model.addAttribute("telematikIdInfo", telematikIdInfo);
-            model.addAttribute("verfuegbar", telematikIdInfo != null);
-            model.addAttribute("suche", true);
+            model.addAttribute("telematikIdInfoVerfuegbar", false);
+            model.addAttribute("holderInfos", new ArrayList<>());
+            model.addAttribute("professionOIDInfos", new ArrayList<>());
         }
+
         return "glossar/glossarUebersicht";
     }
 
     @Operation(description = "Anhand der übergebenen TelematikId werden die Informationen zurückgegeben.",
             responses = { @ApiResponse( responseCode = "200", description = "Anhand der übergebenen TelematikId werden die Informationen zurückgegeben."  )})
-    @RequestMapping(value = "/api/glossar", method = RequestMethod.POST)
+    @RequestMapping(value = "/api/glossar/telematikidinfo", method = RequestMethod.POST)
     @ResponseStatus(value = HttpStatus.OK)
     @ResponseBody
-    public Map<String, TelematikIdInfo> api(@RequestBody List<String> telematikIds) throws Exception {
+    public Map<String, TelematikIdInfo> apiGlossarTelematikIdIndo(@RequestBody List<String> telematikIds) throws Exception {
         Map<String, TelematikIdInfo> res = new HashMap<>();
         for (Iterator<String> iterator = telematikIds.iterator(); iterator.hasNext(); ) {
             String telematikId = iterator.next();
-            res.put(telematikId, glossarService.get(telematikId));
+            res.put(telematikId, glossarService.getTelematikIdInfo(telematikId));
+        }
+        return res;
+    }
+
+    @Operation(description = "Anhand der übergebenen ProfessionOID werden die Informationen zurückgegeben. In dem übergebenen Array können Sie auch nur ein Element vom Wert '*' angeben. Dann bekommen Sie alle verfügbaren Werte.",
+            responses = { @ApiResponse( responseCode = "200", description = "Anhand der übergebenen ProfessionOID werden die Informationen zurückgegeben."  )})
+    @RequestMapping(value = "/api/glossar/professionoidinfo", method = RequestMethod.POST)
+    @ResponseStatus(value = HttpStatus.OK)
+    @ResponseBody
+    public Map<String, ProfessionOIDInfo> apiGlossarProfessionOIDInfo(@RequestBody List<String> professionOIDs) throws Exception {
+        Map<String, ProfessionOIDInfo> res = new HashMap<>();
+
+        if (professionOIDs.size() == 1 && professionOIDs.get(0).equals(GlossarService.SEARCHVALUE_ALL_ELEMS)) {
+            glossarService.getAllProfessionOIDInfo().forEach(professionOIDInfo -> res.put(professionOIDInfo.getCode(), professionOIDInfo));
+        }
+        else {
+            for (Iterator<String> iterator = professionOIDs.iterator(); iterator.hasNext(); ) {
+                String professionOID = iterator.next();
+                res.put(professionOID, glossarService.getProfessionOIDInfo(professionOID));
+            }
+        }
+        return res;
+    }
+
+    @Operation(description = "Anhand des übergebenen Holder werden die Informationen zurückgegeben. In dem übergebenen Array können Sie auch nur ein Element vom Wert '*' angeben. Dann bekommen Sie alle verfügbaren Werte.",
+            responses = { @ApiResponse( responseCode = "200", description = "Anhand des übergebenen Holder werden die Informationen zurückgegeben."  )})
+    @RequestMapping(value = "/api/glossar/holderinfo", method = RequestMethod.POST)
+    @ResponseStatus(value = HttpStatus.OK)
+    @ResponseBody
+    public Map<String, HolderInfo> apiGlossarHolderInfo(@RequestBody List<String> holder) throws Exception {
+        Map<String, HolderInfo> res = new HashMap<>();
+
+        if (holder.size() == 1 && holder.get(0).equals(GlossarService.SEARCHVALUE_ALL_ELEMS)) {
+            glossarService.getAllHolderInfo().forEach(holderInfo -> res.put(holderInfo.getCode(), holderInfo));
+        }
+        else {
+            for (Iterator<String> iterator = holder.iterator(); iterator.hasNext(); ) {
+                String holderV = iterator.next();
+                res.put(holderV, glossarService.getHolderInfo(holderV));
+            }
         }
         return res;
     }