+++ CHANGELOG.md
@@ -13,6 +13,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+## [2.3.0]
+
+### Added
+- Glossar erweitert um ProfessionOID und Holder
+
+### Fixed
+- diverse BugFixes rund um Holder-Attr-Change und Stammdatenimport
+
 ## [2.2.0]
 
 ### Added